constants file in core
