+++ src/binfhe/include/binfhe-constants.h
@@ -32,6 +32,8 @@
 #ifndef _BINFHE_CONSTANTS_H_
 #define _BINFHE_CONSTANTS_H_
 
+#include "lattice/constants-lattice.h"
+
 #include <cstdint>
 #include <iosfwd>
 
@@ -118,14 +120,6 @@ std::ostream& operator<<(std::ostream& s, BINFHE_METHOD f);
 
 enum BINGATE { OR, AND, NOR, NAND, XOR_FAST, XNOR_FAST, MAJORITY, AND3, OR3, AND4, OR4, CMUX, XOR, XNOR };
 std::ostream& operator<<(std::ostream& s, BINGATE f);
-
-/**
- * @brief Lists all modes for secret keys supported
- */
-enum SECRET_KEY_DIST {
-    GAUSSIAN_BINFHE        = 0,
-    UNIFORM_TERNARY_BINFHE = 1,
-};
 
 /**
  * @brief Type of ciphertext generated by the Encrypt method
